+++ src/main.sh
@@ -57,8 +57,8 @@ print_header() {
     print_msg sub-label "- ${STRONG}$LABEL_DISK${NC}: ${DISK_USAGE}"
     print_msg sub-label "- ${STRONG}$LABEL_IPADDR${NC}: ${IP_ADDRESS}"
     echo ""
-    print_msg label "${STRONG}$LABEL_VERSION_CHANNEL${NC}: ${YELLOW}${CORE_CHANNEL}${NC}"
-    core_display_version
+    #print_msg label "${STRONG}$LABEL_VERSION_CHANNEL${NC}: ${YELLOW}${CORE_CHANNEL}${NC}"
+    core_version_display
     echo -e "${MAGENTA}==============================================${NC}"
 }